@@ -3,30 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.1.18",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.26",
-    "@types/react": "^17.0.43",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.5",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
     "@types/react-bootstrap-typeahead": "^5.1.8",
-    "@types/react-dom": "^17.0.14",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.3.3",
-    "@types/sanitize-html": "^2.6.2",
-    "axios": "^0.26.1",
-    "bootstrap": "^4.6.1",
-    "react": "^18.0.0",
-    "react-bootstrap-typeahead": "^5.2.2",
-    "react-dom": "^18.0.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.0",
-    "reactstrap": "^8.10.1",
-    "sanitize-html": "^2.7.0",
-    "typescript": "^4.6.3",
-    "web-vitals": "^2.1.4"
+    "@types/sanitize-html": "^2.8.0",
+    "axios": "^1.2.2",
+    "bootstrap": "^5.2.3",
+    "react": "^18.2.0",
+    "react-bootstrap-typeahead": "^6.0.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.2",
+    "reactstrap": "^9.1.5",
+    "sanitize-html": "^2.8.1",
+    "typescript": "^4.9.4",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -53,6 +52,7 @@
     ]
   },
   "devDependencies": {
-    "prettier": "^2.6.2"
+    "prettier": "^2.8.2",
+    "react-scripts": "5.0.1"
   }
 }
